Made a global static class for functionc, also some more code cleanup
